add django 3.1 classproperty
