+++ django-stubs/utils/functional.pyi
@@ -61,3 +61,9 @@ _PartitionMember = TypeVar("_PartitionMember")
 def partition(
     predicate: Callable, values: List[_PartitionMember]
 ) -> Tuple[List[_PartitionMember], List[_PartitionMember]]: ...
+
+class classproperty:
+    fget: Optional[Callable] = ...
+    def __init__(self, method: Optional[Callable] = ...) -> None: ...
+    def __get__(self, instance: Any, cls: Optional[type] = ...) -> Any: ...
+    def getter(self, method: Callable) -> classproperty: ...
